+++ pytorch_forecasting/layers/_decomposition/__init__.py
@@ -2,7 +2,7 @@
 Decomposition layers for PyTorch Forecasting.
 """
 
-from pytorch_forecasting.layers.decomposition._series_decomp import SeriesDecomposition
+from pytorch_forecasting.layers._decomposition._series_decomp import SeriesDecomposition
 
 __all__ = [
     "SeriesDecomposition",
