+++ pytorch_forecasting/layers/_filter/__init__.py
@@ -2,7 +2,7 @@
 Filtering layers for time series forecasting models.
 """
 
-from pytorch_forecasting.layers.filter._moving_avg_filter import MovingAvg
+from pytorch_forecasting.layers._filter._moving_avg_filter import MovingAvg
 
 __all__ = [
     "MovingAvg",
