+++ pytorch_forecasting/layers/_filter/_moving_avg_filter.py
